chore: adjust RegisterStore

+++ kernel/src/arch/x86_64/utils.rs
@@ -73,8 +73,8 @@ pub unsafe fn get_current_page_table_frame() -> Frame {
     Frame::new((cr3 >> 12) & ((1 << 36) - 1))
 }
 
-#[inline(never)]
-pub unsafe fn bind_pt_and_switch_stack(pt: &X86PageTable, stack: u64, callback: fn() -> !) -> ! {
+#[inline(always)]
+pub unsafe fn bind_pt_and_switch_stack(pt: &X86PageTable, stack: usize, callback: fn() -> !) -> ! {
     unsafe {
         pt.bind();
         asm!(
